+++ src/Products/PloneMeeting/MeetingItem.py
@@ -19,6 +19,7 @@ from DateTime import DateTime
 from datetime import datetime
 from imio.actionspanel.utils import unrestrictedRemoveGivenObject
 from imio.helpers.content import get_vocab
+from imio.helpers.content import safe_delattr
 from imio.helpers.content import uuidsToObjects
 from imio.history.utils import getLastWFAction
 from imio.prettylink.interfaces import IPrettyLink
@@ -96,7 +97,6 @@ from Products.PloneMeeting.utils import display_as_html
 from Products.PloneMeeting.utils import fieldIsEmpty
 from Products.PloneMeeting.utils import forceHTMLContentTypeForEmptyRichFields
 from Products.PloneMeeting.utils import fplog
-from Products.PloneMeeting.utils import get_every_back_references
 from Products.PloneMeeting.utils import getCurrentMeetingObject
 from Products.PloneMeeting.utils import getCustomAdapter
 from Products.PloneMeeting.utils import getFieldVersion
@@ -1330,17 +1330,6 @@ schema = Schema((
             label_msgid='PloneMeeting_label_votesAreSecret',
             i18n_domain='PloneMeeting',
         ),
-    ),
-    ReferenceField(
-        name='predecessor',
-        widget=ReferenceBrowserWidget(
-            visible=False,
-            label='Predecessor',
-            label_msgid='PloneMeeting_label_predecessor',
-            i18n_domain='PloneMeeting',
-        ),
-        multiValued=False,
-        relationship="ItemPredecessor",
     ),
     ReferenceField(
         name='manuallyLinkedItems',
@@ -3802,7 +3791,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
 
     def getUnhandledInheritedAdvisersData(self, adviserUids, optional):
         """ """
-        predecessor = self.getPredecessor()
+        predecessor = self.get_predecessor()
         res = []
         for adviserUid in adviserUids:
             # adviserId could not exist if we removed an inherited initiative advice for example
@@ -4161,7 +4150,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
            If p_checkIsInherited is True, it will check that current advice is actually inherited,
            otherwise, it will not check and return the potential inherited advice."""
         res = None
-        predecessor = self.getPredecessor()
+        predecessor = self.get_predecessor()
         if not predecessor:
             return res
 
@@ -4169,7 +4158,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         while (predecessor and
                predecessor.adviceIndex.get(adviserId) and
                predecessor.adviceIndex[adviserId]['inherited']):
-            predecessor = predecessor.getPredecessor()
+            predecessor = predecessor.get_predecessor()
             inheritedAdviceInfo = deepcopy(predecessor.adviceIndex.get(adviserId))
 
         res = inheritedAdviceInfo
@@ -4956,18 +4945,18 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
 
         # update adviceIndex of every items for which I am the predecessor
         # this way inherited advices are correct if any
-        back_objs = get_every_back_references(self, 'ItemPredecessor')
-        for back_obj in back_objs:
+        successors = self.get_every_successors()
+        for successor in successors:
             # removed inherited advice uids are advice removed on original item
             # that were inherited on back references
             removedInheritedAdviserUids = [
-                adviceInfo['id'] for adviceInfo in back_obj.adviceIndex.values()
+                adviceInfo['id'] for adviceInfo in successor.adviceIndex.values()
                 if adviceInfo.get('inherited', False) and
                 adviceInfo['id'] not in self.adviceIndex]
             if removedInheritedAdviserUids:
                 for removedInheritedAdviserUid in removedInheritedAdviserUids:
-                    back_obj.adviceIndex[removedInheritedAdviserUid]['inherited'] = False
-                back_obj.updateLocalRoles()
+                    successor.adviceIndex[removedInheritedAdviserUid]['inherited'] = False
+                successor.updateLocalRoles()
 
         # notify that advices have been updated so subproducts
         # may interact if necessary
@@ -5516,6 +5505,43 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                        for key, value in sibling.items()}
         return sibling.get(whichItem, sibling)
 
+    def set_predecessor(self, predecessor):
+        '''Only one predecessor possible but several successors.
+           If p_predecessor=None, we remove predecessor/successors attributes.'''
+        if predecessor:
+            self.linked_predecessor_uid = predecessor.UID()
+            if not getattr(predecessor, 'linked_successor_uids', None):
+                predecessor.linked_successor_uids = PersistentList()
+            predecessor.linked_successor_uids.append(self.UID())
+        else:
+            safe_delattr(self, 'linked_predecessor_uid')
+            safe_delattr(self, 'linked_successor_uids')
+
+    def get_predecessor(self, the_object=True, unrestricted=True):
+        ''' '''
+        res = getattr(self, 'linked_predecessor_uid', None)
+        if res and the_object:
+            res = uuidsToObjects(uuids=[res], unrestricted=unrestricted)[0]
+        return res
+
+    def get_successors(self, the_objects=True, unrestricted=True):
+        ''' '''
+        res = getattr(self, 'linked_successor_uids', [])
+        if res and the_objects:
+            # res is a PersistentList, not working with catalog query
+            res = uuidsToObjects(uuids=tuple(res), unrestricted=unrestricted)
+        return res
+
+    def get_every_successors(obj):
+        '''Loop recursievely thru every successors of p_obj and return it.'''
+        def recurse_successors(successors, res=[]):
+            for successor in successors:
+                res.append(successor)
+                recurse_successors(successor.get_successors())
+            return res
+        res = recurse_successors(obj.get_successors())
+        return res
+
     def showDuplicateItemAction_cachekey(method, self, brain=False):
         '''cachekey method for self.showDuplicateItemAction.'''
         return (self, str(self.REQUEST._debug))
@@ -5577,7 +5603,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
            If p_setCurrentAsPredecessor, current item will be set as predecessor
            for the new item, concomitantly if p_manualLinkToPredecessor is True and
            optional field MeetingItem.manuallyLinkedItems is enabled, this will create
-           a manualLink to the predecessor, otherwise, the 'ItemPredecessor' reference is used
+           a manualLink to the predecessor, otherwise, the linked_predecessor_uid is used
            and the link is unbreakable (at least thru the UI).
            If p_inheritAdvices is True, advices will be inherited from predecessor,
            this also needs p_setCurrentAsPredecessor=True and p_manualLinkToPredecessor=False.
@@ -5662,7 +5688,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             if manualLinkToPredecessor:
                 newItem.setManuallyLinkedItems([self.UID()])
             else:
-                newItem.setPredecessor(self)
+                newItem.set_predecessor(self)
                 # manage inherited adviceIds
                 if inheritAdvices:
                     inheritedAdviserUids = [org_uid for org_uid in self.adviceIndex.keys()
@@ -5992,7 +6018,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         if item.meta_type not in ['Plone Site', 'MeetingConfig'] and not item._at_creation_flag:
             # If the item has a predecessor in another meetingConfig we must remove
             # the annotation on the predecessor specifying it.
-            predecessor = self.getPredecessor()
+            predecessor = self.get_predecessor()
             if predecessor:
                 tool = api.portal.get_tool('portal_plonemeeting')
                 cfgId = tool.getMeetingConfig(self).getId()
@@ -6016,12 +6042,12 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
 
     def _cleanAdviceInheritance(self, item, adviceId):
         '''Clean advice inheritance for given p_adviceId on p_item.'''
-        back_objs = get_every_back_references(self, 'ItemPredecessor')
-        for back_obj in back_objs:
-            if back_obj.adviceIndex.get(adviceId, None) and \
-               back_obj.adviceIndex[adviceId]['inherited']:
-                back_obj.adviceIndex[adviceId]['inherited'] = False
-                back_obj.updateLocalRoles()
+        successors = self.get_every_successors()
+        for successor in successors:
+            if successor.adviceIndex.get(adviceId, None) and \
+               successor.adviceIndex[adviceId]['inherited']:
+                successor.adviceIndex[adviceId]['inherited'] = False
+                successor.updateLocalRoles()
 
     security.declarePublic('getAttendees')
 
@@ -6059,30 +6085,26 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                 return False
         return True
 
-    security.declarePublic('getPredecessors')
+    security.declarePublic('get_predecessors')
 
-    def getPredecessors(self, only_viewable=False):
-        '''Returns the list of dict that contains infos about a predecessor.
-           This method can be adapted.'''
+    def get_predecessors(self, only_viewable=False):
+        '''See doc in interfaces.py.'''
         item = self.getSelf()
 
-        predecessor = item.getPredecessor()
+        predecessor = item.get_predecessor()
         predecessors = []
         # retrieve every predecessors
         while predecessor:
             if item._appendLinkedItem(predecessor, only_viewable=only_viewable):
                 predecessors.append(predecessor)
-            predecessor = predecessor.getPredecessor()
+            predecessor = predecessor.get_predecessor()
         # keep order
         predecessors.reverse()
-        # retrieve backrefs too
-        brefs = item.getBRefs('ItemPredecessor')
-        brefs = [bref for bref in brefs if item._appendLinkedItem(bref, only_viewable)]
-        while brefs:
-            predecessors = predecessors + brefs
-            brefs = brefs[0].getBRefs('ItemPredecessor')
-            brefs = [bref for bref in brefs if item._appendLinkedItem(bref, only_viewable)]
-        return predecessors
+        # retrieve successors too
+        successors = item.get_every_successors()
+        successors = [successor for successor in successors
+                      if item._appendLinkedItem(successor, only_viewable)]
+        return predecessors + successors
 
     security.declarePublic('displayLinkedItem')
 
@@ -6375,7 +6397,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
     security.declarePrivate('getAdviceRelatedIndexes')
 
     def getAdviceRelatedIndexes(self):
-        '''See doc in utils.py.'''
+        '''See doc in interfaces.py.'''
         return ['indexAdvisers']
 
     security.declarePublic('lastValidatedBefore')