+++ src/Products/PloneMeeting/tests/testWFAdaptations.py
@@ -2426,8 +2426,8 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.decideMeeting(meeting)
         self.do(item, 'postpone_next_meeting')
         # duplicated and duplicated item is validated
-        clonedItem = item.getBRefs('ItemPredecessor')[0]
-        self.assertEqual(clonedItem.getPredecessor(), item)
+        clonedItem = item.get_successors()[0]
+        self.assertEqual(clonedItem.get_predecessor(), item)
         self.assertEqual(clonedItem.queryState(), 'validated')
         # optional and automatic given advices were inherited
         self.assertTrue(clonedItem.adviceIsInherited(self.vendors_uid))
@@ -2456,8 +2456,8 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.decideMeeting(meeting)
         self.do(item, 'postpone_next_meeting')
         # duplicated and duplicated item is validated
-        clonedItem = item.getBRefs('ItemPredecessor')[0]
-        self.assertEqual(clonedItem.getPredecessor(), item)
+        clonedItem = item.get_successors()[0]
+        self.assertEqual(clonedItem.get_predecessor(), item)
         self.assertEqual(clonedItem.queryState(), 'validated')
 
     def _check_item_decision_state(self,
@@ -2501,10 +2501,10 @@ class testWFAdaptations(PloneMeetingTestCase):
 
         if not will_be_cloned:
             # no predecessor was set
-            self.assertFalse(item.getBRefs('ItemPredecessor'))
+            self.assertEqual(item.successors(the_objects=False), [])
         else:
             # item was duplicated and new item is in it's initial state
-            linked_item = item.getBRefs('ItemPredecessor')[0]
+            linked_item = item.get_successors()[0]
             self.assertEqual(linked_item.queryState(), self._initial_state(linked_item))
 
         # back transition
@@ -2751,6 +2751,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.assertEqual(item.queryState(), 'accepted_out_of_meeting')
         # not duplicated
         self.assertFalse(item.getBRefs())
+        self.assertFalse(item.get_successors())
         # back transition
         self.do(item, 'backToValidatedFromAcceptedOutOfMeeting')
         self.assertEqual(item.queryState(), 'validated')
@@ -2764,8 +2765,8 @@ class testWFAdaptations(PloneMeetingTestCase):
             cfg.setWorkflowAdaptations(wfas)
             cfg.at_post_edit_script()
             self.do(item, 'accept_out_of_meeting')
-            duplicated_item = item.getBRefs()[0]
-            self.assertEqual(duplicated_item.getPredecessor(), item)
+            duplicated_item = item.get_successors()[0]
+            self.assertEqual(duplicated_item.get_predecessor(), item)
             self.assertEqual(duplicated_item.queryState(), 'validated')
             # duplicated_item is not more isAcceptableOutOfMeeting
             self.assertFalse(duplicated_item.getIsAcceptableOutOfMeeting())
@@ -2818,6 +2819,7 @@ class testWFAdaptations(PloneMeetingTestCase):
         self.assertEqual(item.queryState(), 'accepted_out_of_meeting_emergency')
         # not duplicated
         self.assertFalse(item.getBRefs())
+        self.assertEqual(item.get_successors(the_objects=False), [])
         # back transition
         self.do(item, 'backToValidatedFromAcceptedOutOfMeetingEmergency')
         self.assertEqual(item.queryState(), 'validated')
@@ -2831,8 +2833,8 @@ class testWFAdaptations(PloneMeetingTestCase):
             cfg.setWorkflowAdaptations(wfas)
             cfg.at_post_edit_script()
             self.do(item, 'accept_out_of_meeting_emergency')
-            duplicated_item = item.getBRefs()[0]
-            self.assertEqual(duplicated_item.getPredecessor(), item)
+            duplicated_item = item.get_successors()[0]
+            self.assertEqual(duplicated_item.get_predecessor(), item)
             self.assertEqual(duplicated_item.queryState(), 'validated')
             # duplicated_item emergency is no more asked
             self.assertEqual(duplicated_item.getEmergency(), 'no_emergency')