+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changelog
   moreover, make negative `dateGap` work, this is useful to get `Meeting` of today when meeting have no hours defined
 - Make sure the faceted ajax spinner is visible when loading available items on a meeting or page seems somewhat stucked
 - A `MeetingConfig` used in another `MeetingConfig.meetingConfigsToCloneTo` can not be deactivated
+- Removed `MeetingItem.predecessor` `ReferenceField`, manage `predecessor/successors` manually, this will help migrating to DX
 
 4.1.27.2 (2020-06-25)
 ---------------------